requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.2
django-admin==2.0.1
django-cors-headers==3.11.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-heroku==0.3.1
django-six==1.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
psycopg2==2.9.3
psycopg2-binary==2.8.5
PyJWT==2.3.0
pytz==2021.3
requests==2.27.1
screen==1.0.1
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
xlwt==1.3.0
